Make it possible to implement the atomics in config.h
For most cases, the GCC/Intel atomics (__sync functions) are fine,
but there are some systems for which libgcc is incorrectly built
and thus configure reports HAVE_INTEL_ATOMIC_PRIMITIVES
even though the Intel atomics aren't really implemented (i.e. QNX 6.5).

As of now, it is possible to implement the atomics in config.h
and define HB_NO_ATOMIC_IMPL to force HB use *that* implementation.