Add symbols files
This adds the symbols files that are used to export symbols, especially where
a sed-based mechanism is not normally available, where one can control the
symbols that are exported according to the features enabled in the build.

This enables the Visual Studio builds to export .lib files from the DLLs that
are built.

We may be able to use a visibility-based (or __declspec(dllexport) or so)
method for exporting symbols, but since this is within a stable release
series, it might be better to defer it a bit.