Visual Studio builds: Support NDEBUG/HB_NDEBUG/HB_DEBUG
Define HB_NDEBUG by default for release builds, but if desired for further
optimization, allow a new option RELEASE_USE_NDEBUG to the NMake build
files to define NDEBUG during the build to further reduce the amount
of debug items that is built into the binaries, making the binaries more
suitable in production environments.

For debug builds, define HB_DEBUG=1 for enabling debugging code in the
resulting builds, for Visual Studio 2010 or later, as the current code
is not buildable by Visual Studio 2008 with HB_DEBUG=1 defined.